BENTO-15 (Part 3). Adds UUID generation tool and script.
This commit adds a tool and run script that generates a UUID and writes it
to a file in the user's home directory. Various maven pom changes are made
and build resources added to support developing java code in kiji bento. The
release assembly is modified so that the script is properly packaged as part
of bento-cluster in the bento box.